requirements.txt
Django==1.10.3
djangorestframework==3.5.2
pkg-resources==0.0.0
